Final refactoring 	- Move tests to proper folders 	- Create folder for integration tests 	- Add check in ItemController for max URL length
