+++ src/test/java/ControllerTest/HomeControllerTest.java
@@ -4,16 +4,31 @@ import ShopifyProj.Controller.*;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.client.TestRestTemplate;
+import org.springframework.boot.web.server.LocalServerPort;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@SpringBootTest
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class HomeControllerTest {
+
+    @LocalServerPort
+    private int port;
+
+    @Autowired
+    private TestRestTemplate restTemplate;
+
     @Autowired
     private HomeController homeCont;
 
     @Test
     public void homeContextLoads() throws Exception {
         assertThat(homeCont).isNotNull();
+    }
+
+    @Test
+    public void goToHomePage() throws Exception {
+        assertThat(this.restTemplate.getForObject("http://localhost:" + port + "/",
+                String.class)).contains("Home Page");
     }
 }
